# app-principal/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Raiz Conecta API",
    "version": "1.0.0",
    "description": "API da plataforma Raiz Conecta — conecta produtores rurais e mercados. Gerencia autenticação, perfis, demandas, ofertas e administração de usuários."
  },
  "servers": [
    {
      "url": "https://seu-app.vercel.app",
      "description": "Produção (Vercel)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento local"
    }
  ],
...
  "paths": {
    "/api/auth/cadastro": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Cadastrar novo usuário",
        "description": "Cria uma conta de produtor ou mercado. O status inicial é `EM_ANALISE` — o admin precisa aprovar antes do primeiro acesso.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CadastroInput"
              },
              "example": {
                "tipoUsuario": "produtor",
                "nome": "Fazenda Boa Vista",
                "email": "fazenda@email.com",
                "senha": "Senha@123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cadastro realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemSucesso"
                },
                "example": {
                  "message": "Cadastro realizado com sucesso!"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou e-mail já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "examples": {
                  "camposVazios": {
                    "value": {
                      "error": "Preencha todos os campos obrigatórios."
                    }
                  },
                  "emailExistente": {
                    "value": {
                      "error": "Este e-mail já está cadastrado."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErroInterno"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Fazer login",
        "description": "Autentica o usuário com e-mail e senha. Retorna o tipo de usuário para redirecionar ao painel correto.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              },
              "example": {
                "email": "fazenda@email.com",
                "senha": "Senha@123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                },
                "example": {
                  "message": "Login realizado com sucesso!",
                  "tipoUser": "produtor"
                }
              }
            }
          },
          "401": {
            "description": "Senha incorreta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "example": {
                  "error": "Senha incorreta."
                }
              }
            }
          },
          "404": {
            "description": "E-mail não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "example": {
                  "error": "E-mail não encontrado."
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErroInterno"
          }
        }
      }
    },
...
    "/api/admin/usuarios": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Listar todos os usuários",
        "description": "Retorna todos os produtores e mercados cadastrados, ordenados por nome. Cada item inclui o campo `tipo` indicando se é `produtor` ou `mercado`.",
        "responses": {
          "200": {
            "description": "Lista de todos os usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsuarioAdmin"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErroInterno"
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Atualizar status de um usuário",
        "description": "Aprova, rejeita, suspende ou reativa um produtor ou mercado. Atualiza tanto a tabela do perfil quanto a tabela de acesso. Em caso de aprovação ou rejeição, dispara e-mail via microsserviço.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarStatusInput"
              },
              "example": {
                "email": "fazenda@email.com",
                "tipo": "produtor",
                "novoStatus": "APROVADO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemSucesso"
                },
                "example": {
                  "message": "Status alterado para APROVADO"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "example": {
                  "error": "Dados inválidos"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErroInterno"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Excluir usuário permanentemente",
        "description": "Remove o perfil e o acesso de login do usuário. Pode falhar caso existam pedidos vinculados.",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email"
            },
            "example": "fazenda@email.com"
          },
          {
            "name": "tipo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "produtor",
                "mercado"
              ]
            },
            "example": "produtor"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário excluído com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensagemSucesso"
                },
                "example": {
                  "message": "Usuário excluído com sucesso"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "example": {
                  "error": "Dados inválidos"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno — pode haver pedidos vinculados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                },
                "example": {
                  "error": "Erro ao excluir. Pode haver pedidos vinculados."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MensagemSucesso": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Erro": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "CadastroInput": {
        "type": "object",
        "required": [
          "tipoUsuario",
          "nome",
          "email",
          "senha"
        ],
        "properties": {
          "tipoUsuario": {
            "type": "string",
            "enum": [
              "produtor",
              "mercado"
            ]
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "senha": {
            "type": "string",
            "minLength": 6
          }
        }
      },
      "LoginInput": {
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "senha": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "tipoUser": {
            "type": "string",
            "enum": [
              "produtor",
              "mercado",
              "admin"
            ]
          }
        }
      },
...
      "UsuarioAdmin": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "nomeFantasia": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "EM_ANALISE",
              "APROVADO",
              "REJEITADO",
              "SUSPENSO",
              "A_ANALISE"
            ]
          },
          "tipo": {
            "type": "string",
            "enum": [
              "produtor",
              "mercado"
            ]
          },
          "urlDocumento": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AtualizarStatusInput": {
        "type": "object",
        "required": [
          "email",
          "tipo",
          "novoStatus"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "produtor",
              "mercado"
            ]
          },
          "novoStatus": {
            "type": "string",
            "enum": [
              "APROVADO",
              "REJEITADO",
              "SUSPENSO",
              "ATIVO"
            ]
          }
        }
      }
    },
    "responses": {
      "ErroInterno": {
        "description": "Erro interno no servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Erro"
            },
            "example": {
              "error": "Erro interno no servidor"
            }
          }
        }
      }
    }
  }
}